google/stock_analysis_gemini_flash/requirements.txt
streamlit
yfinance
pillow
phidata
duckduckgo-search
google-generativeai
openai